+++ tests/unit/endpoint.spec.ts
@@ -110,7 +110,7 @@ describe("Endpoint", () => {
         options: { inc: 454 },
         logger: loggerMock,
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(0);
+      expect(loggerMock._getLogs().error).toHaveLength(0);
       expect(resultHandlerSpy).toHaveBeenCalledWith({
         error: null,
         input: { n: 453 },
@@ -120,15 +120,13 @@ describe("Endpoint", () => {
         request: requestMock,
         response: responseMock,
       });
-      expect(responseMock.status).toHaveBeenCalledWith(200);
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "success",
-        data: {
-          inc2: 455,
-          str: "453.00",
-          transform: 4,
-        },
-      });
+      expect(responseMock._getStatusCode()).toBe(200);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "success",
+          data: { inc2: 455, str: "453.00", transform: 4 },
+        }),
+      );
     });
 
     test("should close the stream on OPTIONS request", async () => {
@@ -151,28 +149,15 @@ describe("Endpoint", () => {
           }),
         },
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(0);
-      expect(responseMock.status).toHaveBeenCalledWith(200);
-      expect(responseMock.json).toHaveBeenCalledTimes(0);
+      expect(loggerMock._getLogs().error).toHaveLength(0);
+      expect(responseMock._getStatusCode()).toBe(200);
       expect(handlerMock).toHaveBeenCalledTimes(0);
-      expect(responseMock.set).toHaveBeenCalledTimes(4);
-      expect(responseMock.end).toHaveBeenCalledTimes(1);
-      expect(responseMock.set.mock.calls[0]).toEqual([
-        "Access-Control-Allow-Origin",
-        "*",
-      ]);
-      expect(responseMock.set.mock.calls[1]).toEqual([
-        "Access-Control-Allow-Methods",
-        "GET, OPTIONS",
-      ]);
-      expect(responseMock.set.mock.calls[2]).toEqual([
-        "Access-Control-Allow-Headers",
-        "content-type",
-      ]);
-      expect(responseMock.set.mock.calls[3]).toEqual([
-        "X-Custom-Header",
-        "Testing",
-      ]);
+      expect(responseMock._getHeaders()).toEqual({
+        "access-control-allow-origin": "*",
+        "access-control-allow-methods": "GET, OPTIONS",
+        "access-control-allow-headers": "content-type",
+        "x-custom-header": "Testing",
+      });
     });
   });
 
@@ -185,13 +170,13 @@ describe("Endpoint", () => {
         handler: async () => ({ email: "not email" }),
       });
       const { responseMock } = await testEndpoint({ endpoint });
-      expect(responseMock.status).toHaveBeenCalledWith(500);
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "error",
-        error: {
-          message: "output/email: Invalid email",
-        },
-      });
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: { message: "output/email: Invalid email" },
+        }),
+      );
     });
 
     test("Should throw on output parsing non-Zod error", async () => {
@@ -207,14 +192,14 @@ describe("Endpoint", () => {
         }),
       });
       const { responseMock, loggerMock } = await testEndpoint({ endpoint });
-      expect(loggerMock.error).toHaveBeenCalledTimes(1);
-      expect(responseMock.status).toHaveBeenCalledWith(500);
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "error",
-        error: {
-          message: "Something unexpected",
-        },
-      });
+      expect(loggerMock._getLogs().error).toHaveLength(1);
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: { message: "Something unexpected" },
+        }),
+      );
     });
   });
 
@@ -248,16 +233,15 @@ describe("Endpoint", () => {
       });
       expect(handlerMock).toHaveBeenCalledTimes(0);
       expect(middlewareMock).toHaveBeenCalledTimes(1);
-      expect(loggerMock.error).toHaveBeenCalledTimes(0);
-      expect(loggerMock.warn).toHaveBeenCalledTimes(1);
-      expect(loggerMock.warn.mock.calls[0][0]).toBe(
-        "A middleware has closed the stream. Accumulated options:",
-      );
-      expect(loggerMock.warn.mock.calls[0][1]).toEqual({ inc: 454 });
-      expect(responseMock.status).toHaveBeenCalledTimes(0);
-      expect(responseMock.json).toHaveBeenCalledTimes(0);
-      expect(responseMock.statusCode).toBe(200);
-      expect(responseMock.statusMessage).toBe("OK");
+      expect(loggerMock._getLogs().error).toHaveLength(0);
+      expect(loggerMock._getLogs().warn).toEqual([
+        [
+          "A middleware has closed the stream. Accumulated options:",
+          { inc: 454 },
+        ],
+      ]);
+      expect(responseMock._getStatusCode()).toBe(200);
+      expect(responseMock._getStatusMessage()).toBe("OK");
     });
   });
 
@@ -281,10 +265,9 @@ describe("Endpoint", () => {
       const { loggerMock, responseMock, requestMock } = await testEndpoint({
         endpoint,
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(1);
-      expect(loggerMock.error.mock.calls[0][0]).toBe(
-        "Result handler failure: Something unexpected happened.",
-      );
+      expect(loggerMock._getLogs().error).toEqual([
+        ["Result handler failure: Something unexpected happened."],
+      ]);
       expect(spy).toHaveBeenCalledWith({
         error: null,
         logger: loggerMock,
@@ -294,11 +277,8 @@ describe("Endpoint", () => {
         request: requestMock,
         response: responseMock,
       });
-      expect(responseMock.status).toHaveBeenCalledTimes(1);
-      expect(responseMock.status.mock.calls[0][0]).toBe(500);
-      expect(responseMock.json).toHaveBeenCalledTimes(0);
-      expect(responseMock.end).toHaveBeenCalledTimes(1);
-      expect(responseMock.end.mock.calls[0][0]).toBe(
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
         "An error occurred while serving the result: Something unexpected happened.",
       );
     });
@@ -422,12 +402,12 @@ describe("Endpoint", () => {
           body: { n: 123, m: 5 },
         },
       });
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "success",
-        data: {
-          str: "This is fine",
-        },
-      });
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "success",
+          data: { str: "This is fine" },
+        }),
+      );
     });
   });
 
@@ -460,9 +440,12 @@ describe("Endpoint", () => {
           method: "OPTIONS",
         },
       });
-      expect(responseMock.status).toHaveBeenCalledWith(200);
-      expect(responseMock.json).toHaveBeenCalledTimes(0);
-      expect(responseMock.set).toHaveBeenCalledWith("X-Custom-Header", "test");
+      expect(responseMock._getStatusCode()).toBe(200);
+      expect(responseMock._getData()).toBe("");
+      expect(responseMock._getHeaders()).toHaveProperty(
+        "x-custom-header",
+        "test",
+      );
     });
   });
 
@@ -480,14 +463,14 @@ describe("Endpoint", () => {
         }),
       });
       const { responseMock, loggerMock } = await testEndpoint({ endpoint });
-      expect(loggerMock.error).toHaveBeenCalledTimes(1);
-      expect(responseMock.status).toHaveBeenCalledWith(500);
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "error",
-        error: {
-          message: "Something unexpected",
-        },
-      });
+      expect(loggerMock._getLogs().error).toHaveLength(1);
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: { message: "Something unexpected" },
+        }),
+      );
     });
 
     test("thrown in #handleResult()", async () => {
@@ -507,15 +490,11 @@ describe("Endpoint", () => {
         handler: async () => ({ test: "OK" }),
       });
       const { loggerMock, responseMock } = await testEndpoint({ endpoint });
-      expect(loggerMock.error).toHaveBeenCalledTimes(1);
-      expect(loggerMock.error.mock.calls[0][0]).toBe(
-        "Result handler failure: Something unexpected happened.",
-      );
-      expect(responseMock.status).toHaveBeenCalledTimes(1);
-      expect(responseMock.status.mock.calls[0][0]).toBe(500);
-      expect(responseMock.json).toHaveBeenCalledTimes(0);
-      expect(responseMock.end).toHaveBeenCalledTimes(1);
-      expect(responseMock.end.mock.calls[0][0]).toBe(
+      expect(loggerMock._getLogs().error).toEqual([
+        ["Result handler failure: Something unexpected happened."],
+      ]);
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
         "An error occurred while serving the result: Something unexpected happened.",
       );
     });
@@ -538,12 +517,14 @@ describe("Endpoint", () => {
           body: {},
         },
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(1);
-      expect(responseMock.status).toHaveBeenCalledWith(500);
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "error",
-        error: { message: "Something went wrong" },
-      });
+      expect(loggerMock._getLogs().error).toHaveLength(1);
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: { message: "Something went wrong" },
+        }),
+      );
     });
   });
 
@@ -594,11 +575,10 @@ describe("Endpoint", () => {
           },
         },
       });
-      expect(responseMock.json).toHaveBeenCalledWith({
-        data: {},
-        status: "success",
-      });
-      expect(responseMock.status).toHaveBeenCalledWith(200);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({ status: "success", data: {} }),
+      );
+      expect(responseMock._getStatusCode()).toBe(200);
     });
 
     test("should fail during the refinement of invalid inputs", async () => {
@@ -612,13 +592,16 @@ describe("Endpoint", () => {
           },
         },
       });
-      expect(responseMock.json).toHaveBeenCalledWith({
-        error: {
-          message: "dynamicValue: type1Attribute is required if type is type1",
-        },
-        status: "error",
-      });
-      expect(responseMock.status).toHaveBeenCalledWith(400);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: {
+            message:
+              "dynamicValue: type1Attribute is required if type is type1",
+          },
+        }),
+      );
+      expect(responseMock._getStatusCode()).toBe(400);
     });
 
     test("should refine the output schema as well", async () => {
@@ -633,13 +616,13 @@ describe("Endpoint", () => {
           },
         },
       });
-      expect(responseMock.json).toHaveBeenCalledWith({
-        status: "error",
-        error: {
-          message: "output: failure on demand",
-        },
-      });
-      expect(responseMock.status).toHaveBeenCalledWith(500);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: { message: "output: failure on demand" },
+        }),
+      );
+      expect(responseMock._getStatusCode()).toBe(500);
     });
   });
 
@@ -670,11 +653,10 @@ describe("Endpoint", () => {
           },
         },
       });
-      expect(responseMock.json).toHaveBeenCalledWith({
-        data: {},
-        status: "success",
-      });
-      expect(responseMock.status).toHaveBeenCalledWith(200);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({ status: "success", data: {} }),
+      );
+      expect(responseMock._getStatusCode()).toBe(200);
     });
 
     test("should fail during the refinement of invalid inputs", async () => {
@@ -685,13 +667,13 @@ describe("Endpoint", () => {
           body: {},
         },
       });
-      expect(responseMock.json).toHaveBeenCalledWith({
-        error: {
-          message: "Please provide at least one property",
-        },
-        status: "error",
-      });
-      expect(responseMock.status).toHaveBeenCalledWith(400);
+      expect(responseMock._getData()).toBe(
+        JSON.stringify({
+          status: "error",
+          error: { message: "Please provide at least one property" },
+        }),
+      );
+      expect(responseMock._getStatusCode()).toBe(400);
     });
 
     test("should throw when using transformation (constructor)", () => {
@@ -770,11 +752,11 @@ describe("Endpoint", () => {
         },
       });
 
-      expect(loggerMock.debug.mock.calls).toEqual([
+      expect(loggerMock._getLogs().debug).toEqual([
         ["date in mw handler", "object"],
         ["date in endpoint handler", "object"],
       ]);
-      expect(responseMock.status).toHaveBeenCalledWith(200);
+      expect(responseMock._getStatusCode()).toBe(200);
     });
   });
 });