+++ tests/unit/middleware.spec.ts
@@ -1,5 +1,4 @@
 import { z } from "zod";
-import { Request, Response } from "express";
 import { InputValidationError, Middleware } from "../../src";
 import { IOSchemaError } from "../../src/errors";
 import { describe, expect, test, vi } from "vitest";
@@ -65,11 +64,9 @@ describe("Middleware", () => {
         mw.execute({
           input: { test: 123 },
           options: {},
-          logger: makeLoggerMock({ fnMethod: vi.fn }),
-          request: makeRequestMock({ fnMethod: vi.fn }) as unknown as Request,
-          response: makeResponseMock({
-            fnMethod: vi.fn,
-          }) as unknown as Response,
+          logger: makeLoggerMock(),
+          request: makeRequestMock(),
+          response: makeResponseMock(),
         }),
       ).rejects.toThrow(InputValidationError);
     });
@@ -80,16 +77,16 @@ describe("Middleware", () => {
         input: z.object({ test: z.string() }),
         handler: handlerMock,
       });
-      const loggerMock = makeLoggerMock({ fnMethod: vi.fn });
-      const requestMock = makeRequestMock({ fnMethod: vi.fn });
-      const responseMock = makeResponseMock({ fnMethod: vi.fn });
+      const loggerMock = makeLoggerMock();
+      const requestMock = makeRequestMock();
+      const responseMock = makeResponseMock();
       expect(
         await mw.execute({
           input: { test: "something" },
           options: { opt: "anything " },
           logger: loggerMock,
-          request: requestMock as unknown as Request,
-          response: responseMock as unknown as Response,
+          request: requestMock,
+          response: responseMock,
         }),
       ).toEqual({ result: "test" });
       expect(handlerMock).toHaveBeenCalledWith({