+++ tests/unit/result-handler.spec.ts
@@ -1,4 +1,4 @@
-import { Request, Response } from "express";
+import { Response } from "express";
 import createHttpError from "http-errors";
 import { expectType } from "tsd";
 import { z } from "zod";
@@ -68,54 +68,48 @@ describe("ResultHandler", () => {
   });
 
   const requestMock = makeRequestMock({
-    fnMethod: vi.fn,
-    requestProps: { method: "POST", url: "http://something/v1/anything" },
+    method: "POST",
+    url: "http://something/v1/anything",
   });
 
   describe.each([
     {
       subject: defaultResultHandler,
       name: "defaultResultHandler",
-      errorMethod: "json" as const,
     },
     {
       subject: arrayResultHandler,
       name: "arrayResultHandler",
-      errorMethod: "send" as const,
     },
-  ])("$name", ({ subject, errorMethod }) => {
+  ])("$name", ({ subject }) => {
     test("Should handle generic error", () => {
-      const responseMock = makeResponseMock({ fnMethod: vi.fn });
-      const loggerMock = makeLoggerMock({ fnMethod: vi.fn });
+      const responseMock = makeResponseMock();
+      const loggerMock = makeLoggerMock();
       subject.execute({
         error: new Error("Some error"),
         input: { something: 453 },
         output: { anything: 118 },
-        request: requestMock as unknown as Request,
-        response: responseMock as unknown as Response,
+        request: requestMock,
+        response: responseMock,
         logger: loggerMock,
         options: {},
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(1);
-      expect(loggerMock.error.mock.calls[0][0]).toMatch(
-        /^Internal server error\nError: Some error/,
-      );
-      expect(loggerMock.error.mock.calls[0][1]).toHaveProperty("url");
-      expect(loggerMock.error.mock.calls[0][1].url).toBe(
-        "http://something/v1/anything",
-      );
-      expect(loggerMock.error.mock.calls[0][1]).toHaveProperty("payload");
-      expect(loggerMock.error.mock.calls[0][1].payload).toEqual({
-        something: 453,
-      });
-      expect(responseMock.status).toHaveBeenCalledWith(500);
-      expect(responseMock[errorMethod]).toHaveBeenCalledTimes(1);
-      expect(responseMock[errorMethod].mock.calls[0]).toMatchSnapshot();
+      expect(loggerMock._getLogs().error).toEqual([
+        [
+          expect.stringMatching(/^Internal server error\nError: Some error/),
+          {
+            payload: { something: 453 },
+            url: "http://something/v1/anything",
+          },
+        ],
+      ]);
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toMatchSnapshot();
     });
 
     test("Should handle schema error", () => {
-      const responseMock = makeResponseMock({ fnMethod: vi.fn });
-      const loggerMock = makeLoggerMock({ fnMethod: vi.fn });
+      const responseMock = makeResponseMock();
+      const loggerMock = makeLoggerMock();
       subject.execute({
         error: new InputValidationError(
           new z.ZodError([
@@ -131,50 +125,47 @@ describe("ResultHandler", () => {
         input: { something: 453 },
         output: { anything: 118 },
         options: {},
-        request: requestMock as unknown as Request,
-        response: responseMock as unknown as Response,
+        request: requestMock,
+        response: responseMock,
         logger: loggerMock,
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(0);
-      expect(responseMock.status).toHaveBeenCalledWith(400);
-      expect(responseMock[errorMethod]).toHaveBeenCalledTimes(1);
-      expect(responseMock[errorMethod].mock.calls[0]).toMatchSnapshot();
+      expect(loggerMock._getLogs().error).toHaveLength(0);
+      expect(responseMock._getStatusCode()).toBe(400);
+      expect(responseMock._getData()).toMatchSnapshot();
     });
 
     test("Should handle HTTP error", () => {
-      const responseMock = makeResponseMock({ fnMethod: vi.fn });
-      const loggerMock = makeLoggerMock({ fnMethod: vi.fn });
+      const responseMock = makeResponseMock();
+      const loggerMock = makeLoggerMock();
       subject.execute({
         error: createHttpError(404, "Something not found"),
         input: { something: 453 },
         output: { anything: 118 },
         options: {},
-        request: requestMock as unknown as Request,
-        response: responseMock as unknown as Response,
+        request: requestMock,
+        response: responseMock,
         logger: loggerMock,
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(0);
-      expect(responseMock.status).toHaveBeenCalledWith(404);
-      expect(responseMock[errorMethod]).toHaveBeenCalledTimes(1);
-      expect(responseMock[errorMethod].mock.calls[0]).toMatchSnapshot();
+      expect(loggerMock._getLogs().error).toHaveLength(0);
+      expect(responseMock._getStatusCode()).toBe(404);
+      expect(responseMock._getData()).toMatchSnapshot();
     });
 
     test("Should handle regular response", () => {
-      const responseMock = makeResponseMock({ fnMethod: vi.fn });
-      const loggerMock = makeLoggerMock({ fnMethod: vi.fn });
+      const responseMock = makeResponseMock();
+      const loggerMock = makeLoggerMock();
       subject.execute({
         error: null,
         input: { something: 453 },
         output: { anything: 118, items: ["One", "Two", "Three"] },
         options: {},
-        request: requestMock as unknown as Request,
-        response: responseMock as unknown as Response,
+        request: requestMock,
+        response: responseMock,
         logger: loggerMock,
       });
-      expect(loggerMock.error).toHaveBeenCalledTimes(0);
-      expect(responseMock.status).toHaveBeenCalledWith(200);
-      expect(responseMock.json).toHaveBeenCalledTimes(1);
-      expect(responseMock.json.mock.calls[0]).toMatchSnapshot();
+      expect(loggerMock._getLogs().error).toHaveLength(0);
+      expect(responseMock._getStatusCode()).toBe(200);
+      expect(responseMock._getData()).toMatchSnapshot();
     });
 
     test("should forward output schema examples", () => {
@@ -201,20 +192,19 @@ describe("ResultHandler", () => {
   });
 
   test("arrayResultHandler should fail when there is no items prop in the output", () => {
-    const responseMock = makeResponseMock({ fnMethod: vi.fn });
-    const loggerMock = makeLoggerMock({ fnMethod: vi.fn });
+    const responseMock = makeResponseMock();
+    const loggerMock = makeLoggerMock();
     arrayResultHandler.execute({
       error: null,
       input: { something: 453 },
       output: { anything: 118 },
       options: {},
-      request: requestMock as unknown as Request,
-      response: responseMock as unknown as Response,
+      request: requestMock,
+      response: responseMock,
       logger: loggerMock,
     });
-    expect(loggerMock.error).toHaveBeenCalledTimes(0);
-    expect(responseMock.status).toHaveBeenCalledWith(500);
-    expect(responseMock.send).toHaveBeenCalledTimes(1);
-    expect(responseMock.send.mock.calls[0]).toMatchSnapshot();
+    expect(loggerMock._getLogs().error).toHaveLength(0);
+    expect(responseMock._getStatusCode()).toBe(500);
+    expect(responseMock._getData()).toMatchSnapshot();
   });
 });