+++ tests/unit/server-helpers.spec.ts
@@ -22,7 +22,7 @@ import createHttpError from "http-errors";
 describe("Server helpers", () => {
   describe("createParserFailureHandler()", () => {
     test("the handler should call next if there is no error", () => {
-      const rootLogger = makeLoggerMock({ fnMethod: vi.fn });
+      const rootLogger = makeLoggerMock();
       const handler = createParserFailureHandler({
         errorHandler: defaultResultHandler,
         rootLogger,
@@ -44,7 +44,7 @@ describe("Server helpers", () => {
         handler: vi.fn(),
       });
       const spy = vi.spyOn(errorHandler, "execute");
-      const rootLogger = makeLoggerMock({ fnMethod: vi.fn });
+      const rootLogger = makeLoggerMock({ isRoot: true });
       const handler = createParserFailureHandler({
         errorHandler,
         rootLogger,
@@ -53,13 +53,10 @@ describe("Server helpers", () => {
         new SyntaxError("Unexpected end of JSON input"),
         null as unknown as Request,
         makeResponseMock({
-          fnMethod: vi.fn,
-          responseProps: {
-            locals: {
-              [metaSymbol]: { logger: { ...rootLogger, isChild: true } },
-            },
+          locals: {
+            [metaSymbol]: { logger: makeLoggerMock({ isChild: true }) },
           },
-        }) as unknown as Response,
+        }),
         vi.fn<any>(),
       );
       expect(spy).toHaveBeenCalledTimes(1);
@@ -78,33 +75,23 @@ describe("Server helpers", () => {
         handler: vi.fn(),
       });
       const spy = vi.spyOn(errorHandler, "execute");
-      const rootLogger = makeLoggerMock({ fnMethod: vi.fn });
+      const rootLogger = makeLoggerMock({ isRoot: true });
       const handler = createNotFoundHandler({
         errorHandler,
         rootLogger,
       });
       const next = vi.fn();
       const requestMock = makeRequestMock({
-        fnMethod: vi.fn,
-        requestProps: {
-          method: "POST",
-          path: "/v1/test",
-          body: { n: 453 },
-        },
+        method: "POST",
+        path: "/v1/test",
+        body: { n: 453 },
       });
       const responseMock = makeResponseMock({
-        fnMethod: vi.fn,
-        responseProps: {
-          locals: {
-            [metaSymbol]: { logger: { ...rootLogger, isChild: true } },
-          },
+        locals: {
+          [metaSymbol]: { logger: makeLoggerMock({ isChild: true }) },
         },
       });
-      await handler(
-        requestMock as unknown as Request,
-        responseMock as unknown as Response,
-        next,
-      );
+      await handler(requestMock, responseMock, next);
       expect(next).toHaveBeenCalledTimes(0);
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spy.mock.calls[0]).toHaveLength(1);
@@ -120,7 +107,7 @@ describe("Server helpers", () => {
     });
 
     test("should call Last Resort Handler in case of ResultHandler is faulty", () => {
-      const rootLogger = makeLoggerMock({ fnMethod: vi.fn });
+      const rootLogger = makeLoggerMock();
       const errorHandler = new ResultHandler({
         positive: vi.fn(),
         negative: vi.fn(),
@@ -133,25 +120,16 @@ describe("Server helpers", () => {
       });
       const next = vi.fn();
       const requestMock = makeRequestMock({
-        fnMethod: vi.fn,
-        requestProps: {
-          method: "POST",
-          path: "/v1/test",
-          body: { n: 453 },
-        },
+        method: "POST",
+        path: "/v1/test",
+        body: { n: 453 },
       });
-      const responseMock = makeResponseMock({ fnMethod: vi.fn });
-      handler(
-        requestMock as unknown as Request,
-        responseMock as unknown as Response,
-        next,
-      );
+      const responseMock = makeResponseMock();
+      handler(requestMock, responseMock, next);
       expect(next).toHaveBeenCalledTimes(0);
       expect(spy).toHaveBeenCalledTimes(1);
-      expect(responseMock.status).toHaveBeenCalledTimes(1);
-      expect(responseMock.status.mock.calls[0][0]).toBe(500);
-      expect(responseMock.end).toHaveBeenCalledTimes(1);
-      expect(responseMock.end.mock.calls[0][0]).toBe(
+      expect(responseMock._getStatusCode()).toBe(500);
+      expect(responseMock._getData()).toBe(
         "An error occurred while serving the result: I am faulty.\n" +
           "Original error: Can not POST /v1/test.",
       );
@@ -185,19 +163,19 @@ describe("Server helpers", () => {
   });
 
   describe("createUploadLogger()", () => {
-    const rootLogger = makeLoggerMock({ fnMethod: vi.fn });
+    const rootLogger = makeLoggerMock();
     const uploadLogger = createUploadLogger(rootLogger);
 
     test("should debug the messages", () => {
       uploadLogger.log("Express-file-upload: Busboy finished parsing request.");
-      expect(rootLogger.debug).toHaveBeenCalledWith(
-        "Express-file-upload: Busboy finished parsing request.",
-      );
+      expect(rootLogger._getLogs().debug).toEqual([
+        ["Express-file-upload: Busboy finished parsing request."],
+      ]);
     });
   });
 
   describe("createUploadParsers()", async () => {
-    const rootLogger = makeLoggerMock({ fnMethod: vi.fn });
+    const rootLogger = makeLoggerMock();
     const beforeUploadMock = vi.fn();
     const parsers = await createUploadParsers({
       config: {
@@ -214,8 +192,8 @@ describe("Server helpers", () => {
       },
       rootLogger,
     });
-    const requestMock = makeRequestMock({ fnMethod: vi.fn });
-    const responseMock = makeResponseMock({ fnMethod: vi.fn });
+    const requestMock = makeRequestMock();
+    const responseMock = makeResponseMock();
     const nextMock = vi.fn();
 
     test("should return an array of RequestHandler", () => {
@@ -230,22 +208,14 @@ describe("Server helpers", () => {
       beforeUploadMock.mockImplementationOnce(() => {
         throw error;
       });
-      await parsers[0](
-        requestMock as unknown as Request,
-        responseMock as unknown as Response,
-        nextMock,
-      );
+      await parsers[0](requestMock, responseMock, nextMock);
       expect(nextMock).toHaveBeenCalledWith(error);
     });
 
     test("should install the uploader with its special logger", async () => {
       const interalMw = vi.fn();
       fileUploadMock.mockImplementationOnce(() => interalMw);
-      await parsers[0](
-        requestMock as unknown as Request,
-        responseMock as unknown as Response,
-        nextMock,
-      );
+      await parsers[0](requestMock, responseMock, nextMock);
       expect(beforeUploadMock).toHaveBeenCalledWith({
         request: requestMock,
         logger: rootLogger,
@@ -270,14 +240,11 @@ describe("Server helpers", () => {
     test("should place the body into the raw prop of the body object", () => {
       const buffer = Buffer.from([]);
       const requestMock = makeRequestMock({
-        fnMethod: vi.fn,
-        requestProps: {
-          method: "POST",
-          body: buffer,
-        },
+        method: "POST",
+        body: buffer,
       });
       const nextMock = vi.fn();
-      moveRaw(requestMock as unknown as Request, {} as Response, nextMock);
+      moveRaw(requestMock, makeResponseMock(), nextMock);
       expect(requestMock.body).toEqual({ raw: buffer });
       expect(nextMock).toHaveBeenCalled();
     });